Notices: I cleaned up the global pill-shaped notices styles a bit.
They are now slightly more mobile-first with all the breakpoints becoming greater-than.
The narrowest breakpoint has them displayed full screen as banners at the top of the page, edge to edge (like iOS or Android alerts).
The middle breakpoint is at the top right corner of the main content column
The largest breakpoint is also at the top of the main content column, but sticks to the right-edge of the screen.

TODO: need some way to push down the body when a notice is displayed at the mobile size. Perhaps a body-class indicating that a notice is present? Or a placeholder div adjacent to the notice?