Framework: Use an eslintrc with a specific format.
Prior we were using an .eslintrc, which is supposed to be either JSON or YAML. This is now deprecated, in favor of .eslintrc., like .eslintrc.js for JavaScript. See http://eslint.org/docs/user-guide/configuring#using-configuration-files for more details.

This patch moves to an .eslintrc.js and alters the format a bit to export the configuration object as the default, as per the config guide.