Apply width/height style on the Videopress iframe.
The iframe width/height element attribute is ignored because a tinymce css entry overrides it: the VideoPress iframe is a descendant of a div with class wpview-wrap so it gets 100% width. Style on the element has higher priority so it takes effect.

The iframe cannot get too wide because its max-width is 100% as defined by css entry ".wpview-wrap .wpview-content > iframe".