Signup: Add theme preselection flow
We want people to be able to sign up, with a theme already chosen from
an external source, such as the Theme Showcase(s).

We'll want to specify the theme to preselect using an URL, e.g.
`/start/themepick?theme=twentyten`. The selected theme should then be
set as the selected theme, with no additional visual feedback necessary.

- How do we handle premium themes? Can we just add them to the cart?
- Since the preselection of a theme is done externally, how to we fit
  this into the flow? Do we have a dummy step, that selects the
  specified theme, and goes directly to the next step?